Bug Fix
- Dealer backside card isnt displaying and the third card overlaps the second one. Total doesnt match the card image total

+++ src/uicontroller.h
@@ -14,7 +14,8 @@
 #include <qgraphicsitem.h>
 
 
-constexpr int g_card_xOffset = 125;
+//constexpr int g_card_xOffset = 125;
+constexpr uint8_t g_card_xOffset = 125;
 
 
 class UiController
